refactor weight trader
