+++ modules/transaction-payment/src/lib.rs
@@ -36,10 +36,7 @@ use frame_support::{
 		WithdrawReasons,
 	},
 	transactional,
-	weights::{
-		constants::WEIGHT_PER_SECOND, DispatchInfo, GetDispatchInfo, Pays, PostDispatchInfo, WeightToFeeCoefficient,
-		WeightToFeePolynomial,
-	},
+	weights::{DispatchInfo, GetDispatchInfo, Pays, PostDispatchInfo, WeightToFeeCoefficient, WeightToFeePolynomial},
 	BoundedVec, PalletId,
 };
 use frame_system::pallet_prelude::*;
@@ -59,10 +56,8 @@ use sp_runtime::{
 	FixedPointNumber, FixedPointOperand, MultiSignature, Percent, Perquintill,
 };
 use sp_std::prelude::*;
-use support::{DEXManager, PriceProvider, Ratio, SwapLimit, TransactionPayment};
-use xcm::opaque::latest::{prelude::XcmError, AssetId, Fungibility::Fungible, MultiAsset, MultiLocation};
-use xcm_builder::TakeRevenue;
-use xcm_executor::{traits::WeightTrader, Assets};
+use support::{BuyWeightRate, DEXManager, PriceProvider, Rate, Ratio, SwapLimit, TransactionPayment};
+use xcm::opaque::latest::MultiLocation;
 
 mod mock;
 mod tests;
@@ -1074,94 +1069,21 @@ where
 	}
 }
 
-/// `WeightTrader` implementation used for `Trader`, the `rate` is read from storage,
-/// and `token_per_second` is calculated by `rate` * `native_asset_per_second`.
-pub struct TransactionFeePoolTrader<T, C, K: Get<u128>, R: TakeRevenue> {
-	weight: Weight,
-	amount: u128,
-	asset_location: Option<MultiLocation>,
-	asset_per_second: u128,
-	_marker: PhantomData<(T, C, K, R)>,
-}
+pub struct BuyWeightRateOfTransactionFeePool<T, C>(sp_std::marker::PhantomData<(T, C)>);
 
-impl<T: Config, C, K: Get<u128>, R: TakeRevenue> WeightTrader for TransactionFeePoolTrader<T, C, K, R>
+impl<T: Config, C> BuyWeightRate for BuyWeightRateOfTransactionFeePool<T, C>
 where
 	C: Convert<MultiLocation, Option<CurrencyId>>,
 {
-	fn new() -> Self {
-		Self {
-			weight: 0,
-			amount: 0,
-			asset_location: None,
-			asset_per_second: 0,
-			_marker: Default::default(),
-		}
-	}
-
-	fn buy_weight(&mut self, weight: Weight, payment: Assets) -> Result<Assets, XcmError> {
-		// only support first fungible assets now.
-		let asset_id = payment
-			.fungible
-			.iter()
-			.next()
-			.map_or(Err(XcmError::TooExpensive), |v| Ok(v.0))?;
-
-		if let AssetId::Concrete(ref multi_location) = asset_id.clone() {
-			if let Some(token_id) = C::convert(multi_location.clone()) {
-				if let Some(rate) = TokenExchangeRate::<T>::get(token_id) {
-					// calculate the amount of fungible asset.
-					let weight_ratio = Ratio::saturating_from_rational(weight as u128, WEIGHT_PER_SECOND as u128);
-					let asset_per_second = rate.saturating_mul_int(K::get());
-					let amount = weight_ratio.saturating_mul_int(asset_per_second);
-					let required = MultiAsset {
-						id: asset_id.clone(),
-						fun: Fungible(amount),
-					};
-					let unused = payment.checked_sub(required).map_err(|_| XcmError::TooExpensive)?;
-					self.weight = self.weight.saturating_add(weight);
-					self.amount = self.amount.saturating_add(amount);
-					self.asset_location = Some(multi_location.clone());
-					self.asset_per_second = asset_per_second;
-					return Ok(unused);
-				}
-			}
-		}
-		Err(XcmError::TooExpensive)
-	}
-
-	fn refund_weight(&mut self, weight: Weight) -> Option<MultiAsset> {
-		let weight = weight.min(self.weight);
-		let weight_ratio = Ratio::saturating_from_rational(weight as u128, WEIGHT_PER_SECOND as u128);
-		let amount = weight_ratio.saturating_mul_int(self.asset_per_second);
-		self.weight = self.weight.saturating_sub(weight);
-		self.amount = self.amount.saturating_sub(amount);
-		if amount > 0 && self.asset_location.is_some() {
-			Some(
-				(
-					self.asset_location.as_ref().expect("checked is non-empty; qed").clone(),
-					amount,
-				)
-					.into(),
-			)
+	fn calculate_rate(multi_location: MultiLocation) -> Option<Rate> {
+		if let Some(token_id) = C::convert(multi_location) {
+			TokenExchangeRate::<T>::get(token_id)
 		} else {
 			None
 		}
 	}
 }
 
-impl<T, C, K: Get<u128>, R: TakeRevenue> Drop for TransactionFeePoolTrader<T, C, K, R> {
-	fn drop(&mut self) {
-		if self.amount > 0 && self.asset_location.is_some() {
-			R::take_revenue(
-				(
-					self.asset_location.as_ref().expect("checked is non-empty; qed").clone(),
-					self.amount,
-				)
-					.into(),
-			);
-		}
-	}
-}
 impl<T> Convert<Weight, PalletBalanceOf<T>> for Pallet<T>
 where
 	T: Config,
