+++ modules/transaction-payment/src/tests.rs
@@ -1363,39 +1363,6 @@ impl Convert<MultiLocation, Option<CurrencyId>> for CurrencyIdConvert {
 }
 
 #[test]
-fn period_rate_buy_refund_weight_works() {
-	parameter_types! {
-		pub const NativePerSecond: u128 = 8_000_000_000_000;
-	}
-	builder_with_dex_and_fee_pool(true).execute_with(|| {
-		let mock_weight: Weight = 200_000_000;
-		let dot_rate = TokenExchangeRate::<Runtime>::get(DOT);
-		let usd_rate = TokenExchangeRate::<Runtime>::get(AUSD);
-		assert_eq!(dot_rate, Some(Ratio::saturating_from_rational(1, 10)));
-		assert_eq!(usd_rate, Some(Ratio::saturating_from_rational(10, 1)));
-
-		// 1DOT=10KAR, rate=DOT/KAR=1/10, rate=0.1, amount=rate*kar_per_second*weight,
-		// amount=8*weight*rate=0.8*weight=160_000_000
-		let asset: MultiAsset = ((0, X1(GeneralKey(DOT.encode()))), 170_000_000).into();
-		let assets: Assets = asset.into();
-		let mut trader = TransactionFeePoolTrader::<Runtime, CurrencyIdConvert, NativePerSecond, ()>::new();
-		let unused = trader.buy_weight(mock_weight, assets);
-		let expect_asset: MultiAsset = ((0, X1(GeneralKey(DOT.encode()))), 10_000_000).into();
-		assert_eq!(unused.unwrap(), expect_asset.into());
-		assert_eq!(trader.amount, 160_000_000);
-
-		// 1KAR=10AUSD, rate=AUSD/KAR=10, rate=10, amount=8*weight*rate=80*weight=16_000_000_000
-		let asset: MultiAsset = ((0, X1(GeneralKey(AUSD.encode()))), 17_000_000_000).into();
-		let assets: Assets = asset.into();
-		let mut trader = TransactionFeePoolTrader::<Runtime, CurrencyIdConvert, NativePerSecond, ()>::new();
-		let unused = trader.buy_weight(mock_weight, assets);
-		let expect_asset: MultiAsset = ((0, X1(GeneralKey(AUSD.encode()))), 1_000_000_000).into();
-		assert_eq!(unused.unwrap(), expect_asset.into());
-		assert_eq!(trader.amount, 16_000_000_000);
-	});
-}
-
-#[test]
 fn swap_from_pool_not_enough_currency() {
 	builder_with_dex_and_fee_pool(true).execute_with(|| {
 		let balance = 100 as u128;
