+++ runtime/integration-tests/src/setup.rs
@@ -59,7 +59,7 @@ mod mandala_imports {
 		Tokens, TransactionPayment, TransactionPaymentPalletId, TreasuryAccount, TreasuryPalletId, UncheckedExtrinsic,
 		Utility, Vesting, XcmInterface, EVM, NFT,
 	};
-	pub use runtime_common::{cent, dollar, millicent, ACA, AUSD, DOT, KSM, LDOT, LKSM};
+	pub use runtime_common::{cent, dollar, millicent, FixedRateOfAssetRegistry, ACA, AUSD, DOT, KSM, LDOT, LKSM};
 	pub use sp_runtime::traits::AccountIdConversion;
 	use sp_runtime::Percent;
 	pub use xcm_executor::XcmExecutor;
@@ -74,8 +74,7 @@ mod mandala_imports {
 	);
 	pub const NATIVE_TOKEN_SYMBOL: TokenSymbol = TokenSymbol::ACA;
 	pub type Trader = FixedRateOfFungible<DotPerSecond, ()>;
-	pub type PeriodTrader =
-		module_transaction_payment::TransactionFeePoolTrader<Runtime, CurrencyIdConvert, AcaPerSecondAsBased, ()>;
+	pub type TransactionFeePoolTrader = FixedRateOfAssetRegistry<Runtime, CurrencyIdConvert, NativeTokenPerSecond, ()>;
 	pub const ALTERNATIVE_SURPLUS: Percent = AlternativeFeeSurplus::get();
 }
 
@@ -98,8 +97,11 @@ mod karura_imports {
 		TransactionPayment, TransactionPaymentPalletId, TreasuryPalletId, Utility, Vesting, XTokens, XcmInterface, EVM,
 		NFT,
 	};
+	use module_transaction_payment::BuyWeightRateOfTransactionFeePool;
 	pub use primitives::TradingPair;
-	pub use runtime_common::{calculate_asset_ratio, cent, dollar, millicent, KAR, KSM, KUSD, LKSM};
+	pub use runtime_common::{
+		calculate_asset_ratio, cent, dollar, millicent, FixedRateOfAssetRegistry, KAR, KSM, KUSD, LKSM,
+	};
 	pub use sp_runtime::traits::AccountIdConversion;
 	use sp_runtime::Percent;
 	pub use xcm_executor::XcmExecutor;
@@ -124,8 +126,11 @@ mod karura_imports {
 	);
 	pub const NATIVE_TOKEN_SYMBOL: TokenSymbol = TokenSymbol::KAR;
 	pub type Trader = FixedRateOfFungible<KsmPerSecond, ()>;
-	pub type PeriodTrader =
-		module_transaction_payment::TransactionFeePoolTrader<Runtime, CurrencyIdConvert, KarPerSecondAsBased, ()>;
+	pub type TransactionFeePoolTrader = FixedRateOfAssetRegistry<
+		NativeTokenPerSecond,
+		(),
+		BuyWeightRateOfTransactionFeePool<Runtime, CurrencyIdConvert>,
+	>;
 	pub const ALTERNATIVE_SURPLUS: Percent = AlternativeFeeSurplus::get();
 }
 
@@ -148,7 +153,7 @@ mod acala_imports {
 	};
 	pub use frame_support::parameter_types;
 	pub use primitives::TradingPair;
-	pub use runtime_common::{cent, dollar, millicent, ACA, AUSD, DOT, LDOT};
+	pub use runtime_common::{cent, dollar, millicent, FixedRateOfAssetRegistry, ACA, AUSD, DOT, LDOT};
 	pub use sp_runtime::traits::AccountIdConversion;
 	use sp_runtime::Percent;
 	pub use xcm_executor::XcmExecutor;
@@ -175,8 +180,11 @@ mod acala_imports {
 	);
 	pub const NATIVE_TOKEN_SYMBOL: TokenSymbol = TokenSymbol::ACA;
 	pub type Trader = FixedRateOfFungible<DotPerSecond, ()>;
-	pub type PeriodTrader =
-		module_transaction_payment::TransactionFeePoolTrader<Runtime, CurrencyIdConvert, AcaPerSecondAsBased, ()>;
+	pub type TransactionFeePoolTrader = FixedRateOfAssetRegistry<
+		NativeTokenPerSecond,
+		(),
+		BuyWeightRateOfTransactionFeePool<Runtime, CurrencyIdConvert>,
+	>;
 	pub const ALTERNATIVE_SURPLUS: Percent = AlternativeFeeSurplus::get();
 }
